Add details for Tailwind v4 support
Tailwind v4 is now much easier to customize with its css-first configuration and automatic value detection. In this pull request, I simply suggest the addition of a specific section in the readme to explain to users, possibly coming here to upgrade a legacy project, how to add equivalent utilities by hand.